Implemented and tested whitelist for extenders and its constructors.

+++ APITestingCG/test/cz/cvut/fit/hybljan2/apitestingcg/test/TestUtils.java
@@ -33,11 +33,13 @@ public class TestUtils {
     }
 
     public static void delete(File f) throws IOException {
-        if (f.isDirectory()) {
-            for (File c : f.listFiles())
-                delete(c);
+        if(f.exists()) {
+            if (f.isDirectory()) {
+                for (File c : f.listFiles())
+                    delete(c);
+            }
+            if (!f.delete())
+                throw new FileNotFoundException("Failed to delete file: " + f);
         }
-        if (!f.delete())
-            throw new FileNotFoundException("Failed to delete file: " + f);
     }
 }
